[IMP] web: field service: add loadFieldInfo and loadPathDescription
We remove the two hooks useLoadFieldInfo and useLoadPathDescription and
add two corresponding functions loadFieldInfo and loadPathDescription
in the field service.
